Remove some duplicated code adding replStarter and inject context to it
